requirements.txt
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
bcrypt==4.1.2
altair==5.2.0